% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Heatmap_default.R
\name{Heatmap}
\alias{Heatmap}
\title{Generic Heatmap function}
\usage{
Heatmap(x, ...)
}
\arguments{
\item{x}{Object or list of objects of class \code{HP}, \code{DLM}, \code{ClosedHP} or \code{ClosedDLM}. Object of class \code{BLC} or \code{PredBLC}.}

\item{...}{Further arguments passed to or from other methods.}
}
\value{
A ggplot2 heatmap of the life expectancy.
}
\description{
Generic function to \code{Heatmap} method.
}
\seealso{
\code{\link[=Heatmap.HP]{Heatmap.HP()}}, \code{\link[=Heatmap.DLM]{Heatmap.DLM()}}, \code{\link[=Heatmap.BLC]{Heatmap.BLC()}} and \code{\link[=Heatmap.list]{Heatmap.list()}}.
}
