% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_blc.R
\name{print.BLC}
\alias{print.BLC}
\title{BLC: Print}
\usage{
\method{print}{BLC}(x, ...)
}
\arguments{
\item{x}{A \code{BLC} or \code{PredBLC}object, result of a call to blc() or predict() function.}

\item{...}{Further arguments passed to or from other methods.}
}
\value{
A character vector with the details of a fitted \code{BLC} or \code{PredBLC} model.
}
\description{
Print details from a fitted BLC model and returns it invisibly.
}
\seealso{
\code{\link[=print.DLM]{print.DLM()}} and \code{\link[=print.HP]{print.HP()}} for \code{DLM} or \code{HP} methods.
}
