% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_blc_pred.R
\name{print.PredBLC}
\alias{print.PredBLC}
\title{Print Values for BLC prediction models}
\usage{
\method{print}{PredBLC}(x, ...)
}
\arguments{
\item{x}{A \code{PredBLC} object, result to the pred() function call on a \code{BLC} object.}

\item{...}{Further arguments passed to or from other methods.}
}
\value{
A character vector with the details of a fitted \code{PredBLC} model.
}
\description{
Print details from a fitted BLC prediction model and returns it invisibly.
}
\seealso{
\code{\link[=print.DLM]{print.DLM()}}, \code{\link[=print.HP]{print.HP()}} and \code{\link[=print.BLC]{print.BLC()}} for \code{DLM}, \code{HP} or \code{BLC} methods.
}
