% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PT.R
\docType{data}
\name{PT}
\alias{PT}
\title{Mortality Data from Portugal to be used as example}
\format{
A numeric matrix with 63 rows and 16 columns:
\describe{
\item{Row}{Ages available.}
\item{Column}{Years available.}
}
}
\description{
Matrix with the logarithm of the probability of death in Portugal's population from 2000 until 2015.
The ages vary from 18 to 80 years.
}
\references{
Human Mortality Database. University of California, Berkeley (USA), and Max Planck Institute for Demographic Research (Germany). Available at www.mortality.org or www.humanmortality.de (Accessed: July 9th, 2021).
}
\keyword{data}
