% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/USA.R
\docType{data}
\name{USA}
\alias{USA}
\title{Mortality Database from United States to be used as example}
\format{
A data.frame with 9657 rows and 8 variables:
\describe{
\item{Year}{Years available.}
\item{Age}{Ages available.}
\item{Ex.Total}{Exposure of the US population.}
\item{Dx.Total}{Death count of the US population.}
\item{Ex.Male}{Exposure of the US male population.}
\item{Dx.Male}{Death count of the US male population.}
\item{Ex.Female}{Exposure of the US female population.}
\item{Dx.Female}{Death count of the US female population.}
}
}
\description{
Data base with exposures and death counts of the population of the United States between the years 1933 and 2019.
}
\references{
Human Mortality Database. University of California, Berkeley (USA), and Max Planck Institute for Demographic Research (Germany). Available at www.mortality.org or www.humanmortality.de (Accessed: August 9th, 2021).
}
\keyword{data}
