% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_blc.R
\name{print.BLC}
\alias{print.BLC}
\title{Print Values for BLC fitted models}
\usage{
\method{print}{BLC}(x, ...)
}
\arguments{
\item{x}{A \code{BLC} object, result of a call to blc() function.}

\item{...}{Further arguments passed to or from other methods.}
}
\value{
A character vector with the details of a fitted \code{BLC} model.
}
\description{
Print details from a fitted BLC model and returns it invisibly.
}
\seealso{
\code{\link[=print.DLM]{print.DLM()}}, \code{\link[=print.HP]{print.HP()}} and \code{\link[=print.PredBLC]{print.PredBLC()}} for \code{DLM}, \code{HP} or \code{PredBLC} methods.
}
